up to date with flutter's ^3.5.2

+++ lib/src/random.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-double getRandomNumber({int min=0, int max=100}){
+int getRandomNumber({int min=0, int max=100}){
 	Random random = Random();
-	return min.toDouble() + random.nextInt(max-min);
+	return (min.toDouble() + random.nextInt(max-min)).toInt();
 }
